refactor: 🏗️ refactor the architecture in to cleaner patterns
de-spaghettify code written by me
